requirements.txt
alembic==1.16.5
Authlib==1.6.3
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
ConfigArgParse==1.7.1
cryptography==45.0.7
Deprecated==1.2.18
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.4
h11==0.16.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.5.0
locust==2.40.4
locust-cloud==1.26.3
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.1
ordered-set==4.1.0
packaging==25.0
platformdirs==4.4.0
pluggy==1.6.0
psutil==7.0.0
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pyzmq==27.1.0
requests==2.32.5
rich==13.9.4
setuptools==80.9.0
simple-websocket==1.1.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.3
wsproto==1.2.0
zope.event==5.1.1
zope.interface==7.2
